refactor(pkarr): move reqwest-builder test to the tests module
