+++ pkarr/src/client/tests.rs
@@ -720,3 +720,77 @@ async fn discard_cache_with_zero_capacity() {
         "Published packet is not available over the relay only."
     );
 }
+
+#[cfg(feature = "reqwest-builder")]
+mod reqwest_builder {
+    use super::*;
+
+    use std::{
+        net::{SocketAddr, TcpListener},
+        sync::Arc,
+    };
+
+    use axum::routing::get;
+    use axum::Router;
+    use axum_server::tls_rustls::RustlsConfig;
+
+    use crate::{dns::rdata::SVCB, Client, Keypair, SignedPacket};
+
+    async fn publish_server_pkarr(client: &Client, keypair: &Keypair, socket_addr: &SocketAddr) {
+        let mut svcb = SVCB::new(0, ".".try_into().unwrap());
+        svcb.set_port(socket_addr.port());
+
+        let signed_packet = SignedPacket::builder()
+            .https(".".try_into().unwrap(), svcb, 60 * 60)
+            .address(".".try_into().unwrap(), socket_addr.ip(), 60 * 60)
+            .sign(keypair)
+            .unwrap();
+
+        client.publish(&signed_packet, None).await.unwrap();
+    }
+
+    #[rstest]
+    #[case::dht(Networks::Dht)]
+    #[case::both_networks(Networks::Both)]
+    #[cfg_attr(feature = "relays", case::relays(Networks::Relays))]
+    #[tokio::test]
+    async fn reqwest_pkarr_domain(#[case] networks: Networks) {
+        let testnet = mainline::Testnet::new_async(5).await.unwrap();
+        let relay = Relay::run_test(&testnet).await.unwrap();
+
+        let keypair = Keypair::random();
+
+        {
+            // Run a server on Pkarr
+            let app = Router::new().route("/", get(|| async { "Hello, world!" }));
+            let listener = TcpListener::bind("127.0.0.1:0").unwrap(); // Bind to any available port
+            let address = listener.local_addr().unwrap();
+
+            let client = builder(&relay, &testnet, networks).build().unwrap();
+            publish_server_pkarr(&client, &keypair, &address).await;
+
+            println!("Server running on https://{}", keypair.public_key());
+
+            let server = axum_server::from_tcp_rustls(
+                listener,
+                RustlsConfig::from_config(Arc::new((&keypair).into())),
+            );
+
+            tokio::spawn(server.serve(app.into_make_service()));
+        }
+
+        // Client setup
+        let pkarr_client = builder(&relay, &testnet, networks).build().unwrap();
+        let reqwest = reqwest::ClientBuilder::from(pkarr_client).build().unwrap();
+
+        // Make a request
+        let response = reqwest
+            .get(format!("https://{}", keypair.public_key()))
+            .send()
+            .await
+            .unwrap();
+
+        assert_eq!(response.status(), reqwest::StatusCode::OK);
+        assert_eq!(response.text().await.unwrap(), "Hello, world!");
+    }
+}
